requirements.txt
alembic==1.6.5
amqp==5.0.9
aniso8601==9.0.1
APScheduler==3.8.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.6.2
bottle==0.12.19
bottle-websocket==0.2.9
celery==5.2.3
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==41.0.3
dnspython==2.1.0
dominate==2.6.0
EasyExcel==1.1.0
email-validator==1.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Faker==9.8.0
flask==2.3.3
Flask-APScheduler==1.12.3
Flask-Caching==1.10.1
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
flask-marshmallow==0.10.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
Flask-Toastr==0.5.4
Flask-WkHTMLtoPDF==0.1.0
Flask-WTF==1.1.1
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==6.8.0
importlib-resources==5.4.0
inflect==5.3.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.2.1
kombu==5.2.3
lxml==4.7.1
Mako==1.1.4
MarkupSafe==2.1.3
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
MouseInfo==0.1.3
numpy==1.21.4
openpyxl==3.0.9
packaging==23.1
pandas==1.4.0
paramiko==2.9.1
pdfkit==1.0.0
Pillow==8.3.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.1
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.27
pysftp==0.2.9
pytest==7.3.2
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
python-editor==1.0.4
python3-xlib==0.15
PyTweening==1.0.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
text-unidecode==1.3
tomli==2.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
werkzeug==2.3.7
whichcraft==0.6.1
wikipedia==1.4.0
WTForms==2.3.3
XlsxWriter==3.0.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
